requirements.txt
bcrypt==3.2.0
itsdangerous==1.1.0
Flask==1.1.2
python-dotenv==0.15.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Table==0.5.0
Flask-WTF==0.14.3
WTForms==2.3.3
psycopg2==2.8.6
pandas==1.1.4